Rework details of Brics adapter
Previously, determinization was optional in the BricsDFA wrapper (although an
exception was thrown if a non-deterministic Brics automaton was wrapped ín a
BricsDFA without explicitly determinizing it). Now the wrapper always
determinizes the automaton if necessary.

Instead, the boolean parameter is now used (for BricsNFA as well) to optionally
totalize the transition function of the automaton. This allows to conveniently
use Brics automata in structural analysis algorithms (such as equivalence
checks).

This addresses LearnLib/learnlib#60