Remove RichArray and ArrayIterator
Most of the offered functionality was unused and effectively dead code waiting to become a maintenance burden.

Note: the common utility classes are mainly aimed at helping internal development (AutomataLib/LearnLib). We do not intend to provide "general purpose" utilities such as Guava.

Previously, the RichArray class was mainly used to store array-like data with the help of automatic type-inference (e.g. "generic" arrays). To componsate for this usecase, this commits adds an "ArrayStorage" class that is a minimalist wrapper around an Object[] storage, that still allows for a comfortable (generic) handling of different types.